+++ x/ibc/core/02-client/proposal_handler_test.go
@@ -1,0 +1,75 @@
+package client_test
+
+import (
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	distributiontypes "github.com/cosmos/cosmos-sdk/x/distribution/types"
+	govtypes "github.com/cosmos/cosmos-sdk/x/gov/types"
+	client "github.com/cosmos/cosmos-sdk/x/ibc/core/02-client"
+	clienttypes "github.com/cosmos/cosmos-sdk/x/ibc/core/02-client/types"
+	ibctmtypes "github.com/cosmos/cosmos-sdk/x/ibc/light-clients/07-tendermint/types"
+	ibctesting "github.com/cosmos/cosmos-sdk/x/ibc/testing"
+)
+
+func (suite *ClientTestSuite) TestNewClientUpdateProposalHandler() {
+	var (
+		content govtypes.Content
+		err     error
+	)
+
+	testCases := []struct {
+		name     string
+		malleate func()
+		expPass  bool
+	}{
+		{
+			"valid update client proposal", func() {
+				clientA, _ := suite.coordinator.SetupClients(suite.chainA, suite.chainB, ibctesting.Tendermint)
+				clientState := suite.chainA.GetClientState(clientA)
+
+				tmClientState, ok := clientState.(*ibctmtypes.ClientState)
+				suite.Require().True(ok)
+				tmClientState.AllowUpdateAfterMisbehaviour = true
+				tmClientState.FrozenHeight = tmClientState.LatestHeight
+				suite.chainA.App.IBCKeeper.ClientKeeper.SetClientState(suite.chainA.GetContext(), clientA, tmClientState)
+
+				// use next header for chainB to update the client on chainA
+				header, err := suite.chainA.ConstructUpdateTMClientHeader(suite.chainB, clientA)
+				suite.Require().NoError(err)
+
+				content, err = clienttypes.NewClientUpdateProposal(ibctesting.Title, ibctesting.Description, clientA, header)
+				suite.Require().NoError(err)
+			}, true,
+		},
+		{
+			"nil proposal", func() {
+				content = nil
+			}, false,
+		},
+		{
+			"unsupported proposal type", func() {
+				content = distributiontypes.NewCommunityPoolSpendProposal(ibctesting.Title, ibctesting.Description, suite.chainA.SenderAccount.GetAddress(), sdk.NewCoins(sdk.NewCoin("communityfunds", sdk.NewInt(10))))
+			}, false,
+		},
+	}
+
+	for _, tc := range testCases {
+		tc := tc
+
+		suite.Run(tc.name, func() {
+			suite.SetupTest() // reset
+
+			tc.malleate()
+
+			proposalHandler := client.NewClientUpdateProposalHandler(suite.chainA.App.IBCKeeper.ClientKeeper)
+
+			err = proposalHandler(suite.chainA.GetContext(), content)
+
+			if tc.expPass {
+				suite.Require().NoError(err)
+			} else {
+				suite.Require().Error(err)
+			}
+		})
+	}
+
+}
